renaming Palindromes to Palindrome
also removing one implementation, moving the one on root level to the playground to reduce the number of implementations to maintain
